fix warnings for lwcapi
Cleanup minor warnings such as unused references in
lwcapi. Also simplifies to get rid of separate impl
classes. The only api considered public for this
module is the interaction with HTTP interface. The
separate classes were just extra boilerplate.
