+++ atlas-lwcapi/src/main/scala/com/netflix/atlas/lwcapi/ExpressionDatabase.scala
@@ -15,12 +15,86 @@
  */
 package com.netflix.atlas.lwcapi
 
-import com.netflix.atlas.core.model.Query
+import java.util.concurrent.ConcurrentHashMap
+import java.util.concurrent.ScheduledThreadPoolExecutor
+import java.util.concurrent.ThreadFactory
+import java.util.concurrent.TimeUnit
 
-abstract class ExpressionDatabase {
-  def addExpr(expr: ExpressionWithFrequency, query: Query): Boolean
-  def delExpr(id: String): Boolean
-  def hasExpr(id: String): Boolean
-  def expr(id: String): Option[ExpressionWithFrequency]
-  def expressionsForCluster(cluster: String): List[ExpressionWithFrequency]
+import com.netflix.atlas.core.index.QueryIndex
+import com.netflix.atlas.core.model.Query
+import com.netflix.frigga.Names
+import com.typesafe.scalalogging.StrictLogging
+
+import scala.collection.JavaConverters._
+
+class NamedThreadFactory(name: String) extends ThreadFactory {
+  def newThread(r: Runnable): Thread = {
+    val thread = new Thread(r, name)
+    thread.setDaemon(true)
+    thread
+  }
+}
+
+class ExpressionDatabase() extends StrictLogging {
+  import ExpressionDatabase._
+
+  private val knownExpressions = new ConcurrentHashMap[String, Item]().asScala
+  private var queryIndex = QueryIndex.create[ExpressionWithFrequency](Nil)
+
+  @volatile private var queryListChanged = false
+
+  private val ex = new ScheduledThreadPoolExecutor(1, new NamedThreadFactory("ExpressionDatabase"))
+  private val task = new Runnable {
+    def run(): Unit = {
+      regenerateQueryIndex()
+    }
+  }
+  ex.scheduleWithFixedDelay(task, 1, 1, TimeUnit.SECONDS)
+
+  def addExpr(expr: ExpressionWithFrequency, query: Query): Boolean = {
+    // Only replace the object if it is not there, to avoid keeping many identical objects around.
+    val replaced = knownExpressions.putIfAbsent(expr.id, Item(query, expr))
+    val changed = replaced.isEmpty
+    queryListChanged ||= changed
+    changed
+  }
+
+  def delExpr(id: String): Boolean = {
+    val removed = knownExpressions.remove(id)
+    val changed = removed.isDefined
+    queryListChanged |= changed
+    changed
+  }
+
+  def hasExpr(id: String): Boolean = knownExpressions.contains(id)
+
+  def expr(id: String): Option[ExpressionWithFrequency] = {
+    knownExpressions.get(id).map(_.expr)
+  }
+
+  def expressionsForCluster(cluster: String): List[ExpressionWithFrequency] = {
+    val name = Names.parseName(cluster)
+    val tags = Map.newBuilder[String, String]
+    tags += ("nf.cluster" -> name.getCluster)
+    if (name.getApp != null)
+      tags += ("nf.app" -> name.getApp)
+    if (name.getStack != null)
+      tags += ("nf.stack" -> name.getStack)
+    queryIndex.matchingEntries(tags.result)
+  }
+
+  private[lwcapi] def regenerateQueryIndex(): Unit = {
+    if (queryListChanged) {
+      queryListChanged = false
+      val entries = knownExpressions.toList.map { case (_, item) =>
+        QueryIndex.Entry(item.queries, item.expr)
+      }
+      logger.debug(s"Regenerating QueryIndex with ${entries.size} entries")
+      queryIndex = QueryIndex.create(entries)
+    }
+  }
+}
+
+object ExpressionDatabase {
+  case class Item(queries: Query, expr: ExpressionWithFrequency)
 }
