+++ atlas-lwcapi/src/main/scala/com/netflix/atlas/lwcapi/SubscriptionManager.scala
@@ -17,19 +17,55 @@ package com.netflix.atlas.lwcapi
 
 import akka.actor.ActorRef
 
-abstract class SubscriptionManager() {
+import scala.collection.mutable
+
+class SubscriptionManager() {
   import SubscriptionManager._
 
-  def register(streamId: String, ref: ActorRef, name: String): Unit
-  def registration(streamId: String): Option[Entry]
+  private val exprToStream = mutable.Map[String, Set[String]]().withDefaultValue(Set())
+  private val streamToExpr = mutable.Map[String, Set[String]]().withDefaultValue(Set())
+  private val streamToEntry = mutable.Map[String, Entry]()
 
-  def subscribe(streamId: String, expressionId: String): Unit
-  def unsubscribe(streamId: String, expressionId: String): Unit
-  def unregister(streamId: String): List[String]
+  def register(streamId: String, ref: ActorRef, name: String): Unit = synchronized {
+    streamToEntry(streamId) = Entry(streamId, ref, name, System.currentTimeMillis())
+  }
 
-  def actorsForExpression(expressionId: String): Set[ActorRef]
-  def subscribersForExpression(expressionId: String): Set[String]
-  def expressionsForSubscriber(streamId: String): Set[String]
+  def unregister(streamId: String): List[String] = synchronized {
+    streamToEntry.remove(streamId)
+    val ids = streamToExpr.remove(streamId)
+    if (ids.isDefined) {
+      ids.get.foreach(k => exprToStream(k) -= streamId)
+      ids.get.toList
+    } else {
+      List()
+    }
+  }
+
+  def registration(streamId: String): Option[Entry] = synchronized {
+    streamToEntry.get(streamId)
+  }
+
+  def subscribe(streamId: String, expressionId: String): Unit = synchronized {
+    streamToExpr(streamId) += expressionId
+    exprToStream(expressionId) += streamId
+  }
+
+  def unsubscribe(streamId: String, expressionId: String): Unit = synchronized {
+    streamToExpr(streamId) -= expressionId
+    exprToStream(expressionId) -= streamId
+  }
+
+  def actorsForExpression(expressionId: String): Set[ActorRef] = synchronized {
+    exprToStream.getOrElse(expressionId, Set()).flatMap(streamId => streamToEntry.get(streamId)).map(e => e.actorRef)
+  }
+
+  def subscribersForExpression(expressionId: String): Set[String] = synchronized {
+    exprToStream.getOrElse(expressionId, Set())
+  }
+
+  def expressionsForSubscriber(streamId: String): Set[String] = synchronized {
+    streamToExpr.getOrElse(streamId, Set())
+  }
 }
 
 object SubscriptionManager {
